Merge Superdown Files
Originally, the functionality of the game was separated from the menu. Those two projects have now been merged (other than the to-do list). This means I can get to work on the cutscene engine, which currently displays "ENDING [number]" on completion of the game. I also need to add the final two endings (evil hold left ending and true X/O ending), but those can be added later since the cutscenes are going to be a massive sticking point for me.