Created cutscene proof
FINALLY! It cycles through each set of 3 bars individually, and adds a rectangle on top to block out the text, creating the illusion of a typewriter effect with none of the work. Will still need to add images and rewrite the scripts, as well as creating an actual "the end" screen.