Improved cutscene dialogue
Created a function to automatically split the walls of text, and made both cutscene writers sync up with their respective scenes. Also touched up some of the dialogue for the hold ending. This is being saved early so I can work on potentially dismantling the jumpscare _update() function and rebuilding it with flip(), which should make the rest of that cutscene easier to implement.