Push to A30
Making sure the current build is pushed, then exporting it as a png so I can test how it feels on my Miyoo A30 (which I am so hype to have finally). Hopefully the next contribution should be to the normal ending (i.e. music, picture, the end screen, and cutscene writing) so I can start showing it off to people.